helper,activedecoratorを使ってviewのメソッドをモデルから移動
