+++ app/models/user.rb
@@ -44,15 +44,7 @@ class User < ApplicationRecord
     likes.exists?(article_id: article.id)
   end
 
-  def display_name
-    # if profile && profile.nickname
-    #   profile.nickname
-    # else
-    #   self.email.split('@').first
-    # end
-    #ぼっち演算子
-    profile&.nickname || self.email.split('@').first
-  end
+  
 
   def follow!(user)
     user_id = get_user_id(user)
@@ -73,13 +65,7 @@ class User < ApplicationRecord
     profile || build_profile
   end
 
-  def avatar_image
-    if profile&.avatar&.attached?
-      profile.avatar
-    else
-      'default-avatar.png'
-    end
-  end
+  
 
   private
   def get_user_id(user)
